+++ src/acf.c
@@ -228,7 +228,7 @@ SEXP R_flint_acf_initialize(SEXP object, SEXP s_length, SEXP s_x,
 		ny = asLength(s_length, __func__);
 	else
 		ny = 0;
-	acf_ptr y = (acf_ptr) ((ny) ? flint_calloc((size_t) ny, sizeof(acf_t)) : 0);
+	acf_ptr y = (acf_ptr) ((ny) ? flint_calloc(ny, sizeof(acf_t)) : 0);
 	R_flint_set(object, y, ny, (R_CFinalizer_t) &R_flint_acf_finalize);
 	if (s_real != R_NilValue || s_imag != R_NilValue) {
 		if (s_real != R_NilValue) {
@@ -461,7 +461,7 @@ SEXP R_flint_acf_part(SEXP object, SEXP s_op)
 	acf_srcptr x = (acf_ptr) R_flint_get_pointer(object);
 	int op = INTEGER_RO(s_op)[0];
 	SEXP ans = PROTECT(newObject("arf"));
-	arf_ptr y = (arf_ptr) ((n) ? flint_calloc((size_t) n, sizeof(arf_t)) : 0);
+	arf_ptr y = (arf_ptr) ((n) ? flint_calloc(n, sizeof(arf_t)) : 0);
 	R_flint_set(ans, y, n, (R_CFinalizer_t) &R_flint_arf_finalize);
 	if (op == 0)
 	for (j = 0; j < n; ++j)
@@ -553,7 +553,7 @@ SEXP R_flint_acf_ops2(SEXP s_op, SEXP s_x, SEXP s_y)
 #endif
 	{
 		SEXP ans = newObject("acf");
-		acf_ptr z = (acf_ptr) ((n) ? flint_calloc((size_t) n, sizeof(acf_t)) : 0);
+		acf_ptr z = (acf_ptr) ((n) ? flint_calloc(n, sizeof(acf_t)) : 0);
 		R_flint_set(ans, z, n, (R_CFinalizer_t) &R_flint_acf_finalize);
 		switch (op) {
 		case 1: /*   "+" */
@@ -666,7 +666,7 @@ SEXP R_flint_acf_ops1(SEXP s_op, SEXP s_x, SEXP s_dots)
 	case 49: /*  "signif" */
 	{
 		SEXP ans = newObject("acf");
-		acf_ptr z = (acf_ptr) ((n) ? flint_calloc((size_t) n, sizeof(acf_t)) : 0);
+		acf_ptr z = (acf_ptr) ((n) ? flint_calloc(n, sizeof(acf_t)) : 0);
 		R_flint_set(ans, z, n, (R_CFinalizer_t) &R_flint_acf_finalize);
 		switch (op) {
 		case  1: /*       "+" */
@@ -1001,7 +1001,7 @@ SEXP R_flint_acf_ops1(SEXP s_op, SEXP s_x, SEXP s_dots)
 	case 13: /*      "abs" */
 	{
 		SEXP ans = newObject("arf");
-		arf_ptr z = (arf_ptr) ((n) ? flint_calloc((size_t) n, sizeof(arf_t)) : 0);
+		arf_ptr z = (arf_ptr) ((n) ? flint_calloc(n, sizeof(arf_t)) : 0);
 		R_flint_set(ans, z, n, (R_CFinalizer_t) &R_flint_arf_finalize);
 		switch (op) {
 		case  9: /*       "Re" */
